feat: 65% de las funcionalidades entrega 2

+++ app/src/main/java/com/example/postmatch/ui/Screens/perfil/PerfilScreen.kt
@@ -31,7 +31,6 @@ import androidx.compose.ui.Modifier
 import androidx.compose.material3.Icon
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.res.colorResource
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.foundation.lazy.items
@@ -48,19 +47,14 @@ import androidx.compose.runtime.getValue
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import androidx.lifecycle.viewmodel.compose.viewModel
 import coil.compose.AsyncImage
 import coil.request.ImageRequest
 import com.example.postmatch.R
 import com.example.postmatch.data.ReviewInfo
 
 
-data class ReseniaPerfilData(val nEstrellas: Int, val tituloReseniaPerfil: String, val descripcionReseniaPerfil: String, val idFoto: Int)
-
-//@Preview(showBackground = true)
 @Composable
 fun PerfilScreen(
     perfilViewModel: PerfilViewModel,
@@ -68,14 +62,12 @@ fun PerfilScreen(
     reviewButtonClick: () -> Unit,
     idPerfilUsuario: String,
     onReviewClick: (String) -> Unit,
-    //idPartido: Int,
     modifier: Modifier = Modifier
 ){
 
     LaunchedEffect(Unit) {
         perfilViewModel.getUserInfo(idPerfilUsuario)
     }
-
 
     val state by perfilViewModel.uiState.collectAsState()
     LazyColumn(
@@ -92,7 +84,7 @@ fun PerfilScreen(
                 fotoPerfilUrl = state.usuarioInfo.fotoPerfil,
                 nombrePerfil = state.usuarioInfo.nombre,
                 arrobaPerfil = state.usuarioInfo.email,
-                onFotoPerfilButton = {},//perfilViewModel::uploadProfileImageToFirebase,
+                onFotoPerfilButton = perfilViewModel::uploadProfileImageToFirebase,
                 oficioPerfil = "Futbolista")
         }
         item {
@@ -101,17 +93,18 @@ fun PerfilScreen(
         item {
             TextoIzquierda(stringResource(R.string.rese_as))
         }
-        items(state.reviews) { resenhia ->
+        items(state.reviews) { review ->
             ItemReseniaPerfil(
-                reseniaPerfil = resenhia,
-                onDeleteReview = { perfilViewModel.onDeleteReview(resenhia.idReview) }, // eliminar
+                reseniaPerfil = review,
+                onDeleteReview = { perfilViewModel.onDeleteReview(review.idReview) }, // eliminar
                 onClickReview = { idReview ->
                     // Acción para ver o seleccionar la reseña
                 },
                 onReviewClick = { idReview ->
                     // Acción para editar la reseña
                     onReviewClick(idReview) // callback que viene de PerfilScreen
-                }
+                },
+                isCurrentUser = state.isCurrentUser
             )
         }
 
@@ -120,26 +113,12 @@ fun PerfilScreen(
 }
 
 @Composable
-fun SeccionReseniasPerfil(
-    modifier: Modifier = Modifier,
-    resenhias: List<ReviewInfo> // Define el tipo correctamente aquí
-) {
-    // Usamos LazyColumn para listas dinámicas y de mayor rendimiento
-    LazyColumn(
-        modifier = modifier
-            .background(colorResource(id = R.color.verde_oscuro))
-            .padding(vertical = 8.dp)
-    ) {
-    }
-}
-
-@Composable
 fun ItemReseniaPerfil(
     reseniaPerfil: ReviewInfo,
     onDeleteReview: (String) -> Unit,
     onClickReview: (String) -> Unit,
     onReviewClick: (String) -> Unit,
-
+    isCurrentUser: Boolean,
     modifier: Modifier = Modifier
 ) {
     Row(
@@ -190,26 +169,28 @@ fun ItemReseniaPerfil(
                 horizontalArrangement = Arrangement.Start,
                 verticalAlignment = Alignment.CenterVertically
             ) {
-                IconButton(
-                    onClick = { onReviewClick(reseniaPerfil.idReview) },
-                    modifier = Modifier.size(24.dp)
-                ) {
-                    Icon(
-                        imageVector = Icons.Default.Edit,
-                        contentDescription = "Modificar reseña",
-                        tint = MaterialTheme.colorScheme.primary
-                    )
-                }
+                if(isCurrentUser) {
+                    IconButton(
+                        onClick = { onReviewClick(reseniaPerfil.idReview) },
+                        modifier = Modifier.size(24.dp)
+                    ) {
+                        Icon(
+                            imageVector = Icons.Default.Edit,
+                            contentDescription = "Modificar reseña",
+                            tint = MaterialTheme.colorScheme.primary
+                        )
+                    }
 
-                IconButton(
-                    onClick = { onDeleteReview(reseniaPerfil.idReview) },
-                    modifier = Modifier.size(24.dp)
-                ) {
-                    Icon(
-                        imageVector = Icons.Default.Delete,
-                        contentDescription = "Eliminar reseña",
-                        tint = MaterialTheme.colorScheme.error
-                    )
+                    IconButton(
+                        onClick = { onDeleteReview(reseniaPerfil.idReview) },
+                        modifier = Modifier.size(24.dp)
+                    ) {
+                        Icon(
+                            imageVector = Icons.Default.Delete,
+                            contentDescription = "Eliminar reseña",
+                            tint = MaterialTheme.colorScheme.error
+                        )
+                    }
                 }
             }
         }
@@ -446,20 +427,3 @@ fun PickImageButton(
         Text(text = "Seleccionar Imagen")
     }
 }
-/*
-@Composable
-//@Preview(showBackground = true)
-fun PerfilScreenPreview(
-    modifier : Modifier = Modifier
-){
-    //ImagenPerfil(fotoPerfil = R.drawable.ricardo_icon, nombrePerfil = "Ricardito", arrobaPerfil = "@Ricardo_420", oficioPerfil = "Futbolista")
-    PerfilScreen(
-        configuracionButtonClick = {},
-        reviewButtonClick = {},
-        perfilViewModel = viewModel(),
-        idPerfilUsuario = 1
-    )
-}
-
-
- */
